Temporary changes for Autonomous.
