+++ src/main/java/com/synopsys/integration/detect/util/filter/DetectToolFilter.java
@@ -39,7 +39,6 @@ public class DetectToolFilter {
         RunDecision runDecision,
         BlackDuckDecision blackDuckDecision
     ) {
-
         this.excludedIncludedFilter = excludedIncludedFilter;
         this.impactEnabled = impactEnabled;
         this.iacEnabled = iacEnabled;
